\name{MultiCOAP-package}
\alias{MultiCOAP-package}
\alias{MultiCOAP}
\docType{package}
\title{
\packageTitle{MultiCOAP}
}
\description{
\packageDescription{MultiCOAP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MultiCOAP}
\packageIndices{MultiCOAP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MultiCOAP}

Maintainer: \packageMaintainer{MultiCOAP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
